Polyfills for PhoneGap plugins.
Polyfills for plugins provided through phonegap.js. The polyfills will
prevent runtime errors when running an HTML5 app locally with
phonegap.js referenced. This script is mainly intended for development
using PhoneGap build. A polyfill should be added for each referenced
plugin. This script should not be included in the package submitted to
build.phonegap.com.